requirements.txt
django==3.0
djangorestframework==3.11.0
gunicorn==20.0.4
django-background-task==0.1.8

nltk==3.4.5
pdfminer.six==20200402
numpy==1.18.3
tqdm==4.45.0
seaborn==0.10.1
matplotlib==3.2.1
pandas==1.0.3